requirements.txt
flask 
flask-cors
firebase-admin 
python-dotenv 
requests
transformers 
sentencepiece 
torch